fixup! feat(core): Mark components for check if they read a signal
