+++ packages/core/src/render3/instructions/listener.ts
@@ -17,8 +17,9 @@ import {CLEANUP, CONTEXT, LView, RENDERER, TView} from '../interfaces/view';
 import {assertTNodeType} from '../node_assert';
 import {profiler, ProfilerEvent} from '../profiler';
 import {getCurrentDirectiveDef, getCurrentTNode, getLView, getTView} from '../state';
-import {getComponentLViewByIndex, getNativeByTNode, markViewDirty, unwrapRNode} from '../util/view_utils';
+import {getComponentLViewByIndex, getNativeByTNode, unwrapRNode} from '../util/view_utils';
 
+import {markViewDirty} from './mark_view_dirty';
 import {getOrCreateLViewCleanup, getOrCreateTViewCleanup, handleError, loadComponentRenderer} from './shared';
