+++ packages/core/src/render3/util/change_detection_utils.ts
@@ -9,9 +9,9 @@
 import {assertDefined} from '../../util/assert';
 import {getComponentViewByInstance} from '../context_discovery';
 import {detectChanges} from '../instructions/change_detection';
+import {markViewDirty} from '../instructions/mark_view_dirty';
 
 import {getRootComponents} from './discovery_utils';
-import {markViewDirty} from './view_utils';
 
 /**
  * Marks a component for check (in case of OnPush components) and synchronously
