+++ packages/core/src/signals/src/watch.ts
@@ -6,8 +6,7 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {Consumer, consumerPollValueStatus, Edge, nextReactiveId, ProducerId, setActiveConsumer} from './graph';
-import {newWeakRef} from './weak_ref';
+import {BaseConsumer, consumerPollValueStatus, setActiveConsumer} from './graph';
 
 /**
  * Watches a reactive expression and allows it to be scheduled to re-run
@@ -16,17 +15,14 @@ import {newWeakRef} from './weak_ref';
  * `Watch` doesn't run reactive expressions itself, but relies on a consumer-
  * provided scheduling operation to coordinate calling `Watch.run()`.
  */
-export class Watch implements Consumer {
-  readonly id = nextReactiveId();
-  readonly ref = newWeakRef(this);
-  readonly producers = new Map<ProducerId, Edge>();
-  trackingVersion = 0;
-
+export class Watch extends BaseConsumer {
   private dirty = false;
 
-  constructor(private watch: () => void, private schedule: (watch: Watch) => void) {}
+  constructor(private watch: () => void, private schedule: (watch: Watch) => void) {
+    super();
+  }
 
-  notify(): void {
+  override notify(): void {
     if (!this.dirty) {
       this.schedule(this);
     }
