+++ packages/core/test/acceptance/change_detection_signals_in_zones_spec.ts
@@ -9,7 +9,6 @@
 import {ChangeDetectionStrategy, Component} from '@angular/core';
 import {TestBed} from '@angular/core/testing';
 
-import {getReactiveLViewConsumer} from '../../src/render3/reactive_lview_consumer';
 import {signal} from '../../src/signals';
 import {nextReactiveId} from '../../src/signals/src/graph';
 
@@ -58,167 +57,5 @@ describe('CheckAlways components', () => {
     value.set('new');
     fixture.detectChanges();
     expect(fixture.nativeElement.textContent.trim()).toBe('new');
-  });
-});
-
-describe('OnPush components with signals', () => {
-  it('marks view dirty', () => {
-    @Component({
-      template: `{{value()}}{{incrementTemplateExecutions()}}`,
-      changeDetection: ChangeDetectionStrategy.OnPush,
-      standalone: true,
-    })
-    class OnPushCmp {
-      numTemplateExecutions = 0;
-      value = signal('initial');
-      incrementTemplateExecutions() {
-        this.numTemplateExecutions++;
-        return '';
-      }
-    }
-    const fixture = TestBed.createComponent(OnPushCmp);
-    const instance = fixture.componentInstance;
-
-    fixture.detectChanges();
-    expect(instance.numTemplateExecutions).toBe(1);
-    expect(fixture.nativeElement.textContent.trim()).toEqual('initial');
-
-    fixture.detectChanges();
-    // Should not be dirty, should not execute template
-    expect(instance.numTemplateExecutions).toBe(1);
-
-    fixture.componentInstance.value.set('new');
-    fixture.detectChanges();
-    expect(instance.numTemplateExecutions).toBe(2);
-    expect(instance.value()).toBe('new');
-  });
-
-  it('can read a signal in a host binding', () => {
-    @Component({
-      template: `{{incrementTemplateExecutions()}}`,
-      selector: 'child',
-      host: {'[class.blue]': 'useBlue()'},
-      changeDetection: ChangeDetectionStrategy.OnPush,
-      standalone: true,
-    })
-    class ChildCmp {
-      useBlue = signal(false);
-
-      numTemplateExecutions = 0;
-      incrementTemplateExecutions() {
-        this.numTemplateExecutions++;
-        return '';
-      }
-    }
-
-    @Component({
-      template: `<child />`,
-      changeDetection: ChangeDetectionStrategy.OnPush,
-      imports: [ChildCmp],
-      standalone: true,
-    })
-    class ParentCmp {
-    }
-    const fixture = TestBed.createComponent(ParentCmp);
-    const child = fixture.debugElement.query(p => p.componentInstance instanceof ChildCmp);
-    const childInstance = child.componentInstance as ChildCmp;
-
-    fixture.detectChanges();
-    expect(childInstance.numTemplateExecutions).toBe(1);
-    expect(child.nativeElement.outerHTML).not.toContain('blue');
-
-    childInstance.useBlue.set(true);
-    fixture.detectChanges();
-    // We should not re-execute the child template. It didn't change, the host bindings did.
-    expect(childInstance.numTemplateExecutions).toBe(1);
-    expect(child.nativeElement.outerHTML).toContain('blue');
-  });
-
-  it('can have signals in both template and host bindings', () => {
-    @Component({
-      template: ``,
-      selector: 'child',
-      host: {'[class.blue]': 'useBlue()'},
-      changeDetection: ChangeDetectionStrategy.OnPush,
-      standalone: true,
-    })
-    class ChildCmp {
-      useBlue = signal(false);
-    }
-
-    @Component({
-      template: `<child /> {{parentSignalValue()}}`,
-      changeDetection: ChangeDetectionStrategy.OnPush,
-      imports: [ChildCmp],
-      standalone: true,
-      selector: 'parent',
-    })
-    class ParentCmp {
-      parentSignalValue = signal('initial');
-    }
-
-    // Wrapper component so we can effectively test ParentCmp being marked dirty
-    @Component({
-      template: `<parent />`,
-      changeDetection: ChangeDetectionStrategy.OnPush,
-      imports: [ParentCmp],
-      standalone: true,
-    })
-    class TestWrapper {
-    }
-
-    const fixture = TestBed.createComponent(TestWrapper);
-    const parent = fixture.debugElement.query(p => p.componentInstance instanceof ParentCmp)
-                       .componentInstance as ParentCmp;
-    const child = fixture.debugElement.query(p => p.componentInstance instanceof ChildCmp)
-                      .componentInstance as ChildCmp;
-
-    fixture.detectChanges();
-    expect(fixture.nativeElement.outerHTML).toContain('initial');
-    expect(fixture.nativeElement.outerHTML).not.toContain('blue');
-
-    child.useBlue.set(true);
-    fixture.detectChanges();
-    expect(fixture.nativeElement.outerHTML).toContain('blue');
-
-    // Set the signal in the parent again and ensure it gets updated
-    parent.parentSignalValue.set('new');
-    fixture.detectChanges();
-    expect(fixture.nativeElement.outerHTML).toContain('new');
-
-    // Set the signal in the child host binding again and ensure it is still updated
-    child.useBlue.set(false);
-    fixture.detectChanges();
-    expect(fixture.nativeElement.outerHTML).not.toContain('blue');
-  });
-
-  it('does not generate a reactive consumer for each view', () => {
-    @Component({
-      template: ``,
-      selector: 'child',
-      standalone: true,
-    })
-    class ChildCmp {
-    }
-
-    @Component({
-      template: `<child /> <child /> <child />`,
-      imports: [ChildCmp],
-      standalone: true,
-    })
-    class ParentCmp {
-    }
-
-    const reactiveIdBeforeViewCreation = nextReactiveId();
-    TestBed.createComponent(ParentCmp);
-    const reactiveIdAfterViewCreation = nextReactiveId();
-
-    // Only one reactive ID should get generated even though we have at least 4 views (1 parent and
-    // 3 children).
-    // We would ordinarily expect this to be one. However, it might be zero instead, depending on
-    // the order in which the tests are run. This is because the whole point of the optimization is
-    // to lazily assign the next reactive consumer, so there might be one already allocated in a
-    // previous test, waiting to be claimed.
-    expect(reactiveIdAfterViewCreation - reactiveIdBeforeViewCreation).toBeLessThanOrEqual(1);
   });
 });
