+++ packages/core/test/bundling/forms_reactive/bundle.golden_symbols.json
@@ -48,6 +48,9 @@
     "name": "BROWSER_MODULE_PROVIDERS_MARKER"
   },
   {
+    "name": "BaseConsumer"
+  },
+  {
     "name": "BaseControlValueAccessor"
   },
   {
@@ -1054,6 +1057,9 @@
   },
   {
     "name": "getOrCreateComponentTView"
+  },
+  {
+    "name": "getOrCreateCurrentLViewConsumer"
   },
   {
     "name": "getOrCreateInjectable"
