+++ packages/core/test/bundling/forms_template_driven/bundle.golden_symbols.json
@@ -51,6 +51,9 @@
     "name": "BROWSER_MODULE_PROVIDERS_MARKER"
   },
   {
+    "name": "BaseConsumer"
+  },
+  {
     "name": "BaseControlValueAccessor"
   },
   {
@@ -1015,6 +1018,9 @@
   },
   {
     "name": "getOrCreateComponentTView"
+  },
+  {
+    "name": "getOrCreateCurrentLViewConsumer"
   },
   {
     "name": "getOrCreateInjectable"
